# backend/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateShareInput:
    description: Create share link request body
    properties:
      duration_minute:
        minimum: 1
        type: integer
      media_uuid:
        type: string
      target_user_id:
        type: integer
    required:
    - duration_minute
    - media_uuid
    type: object
...
  response.ApiResponse:
    properties:
      code:
        description: 0 success, non-zero failure
        type: integer
      data:
        description: Optional payload
      message:
        description: Human-readable message
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: PrismBox backend HTTP API reference
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: PrismBox Backend API
  version: "1.0"
paths:
...
  /s/{share_token}:
    get:
      description: Public HTML viewer for shared media
      parameters:
      - description: Share token
        in: path
        name: share_token
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML page with embedded media
        "400":
          description: Share not found, revoked, or expired
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: View shared resource
      tags:
      - Shares
...
  /shares:
    post:
      consumes:
      - application/json
      description: Share with optional target user or public link (authentication
        required)
      parameters:
      - description: Share payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateShareInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK; returns share URL
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Media not found, forbidden, or target user missing
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create share link
      tags:
      - Shares
  /shares/{share_token}/meta:
    get:
      description: Public; no authentication required
      parameters:
      - description: Share token
        in: path
        name: share_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Share not found, revoked, or expired
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Get share metadata
      tags:
      - Shares
  /shares/with-me:
    get:
      description: Items shared with the current user (authentication required)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: List shares with me
      tags:
      - Shares
...
securityDefinitions:
  BearerAuth:
    description: 'Use `Authorization: Bearer {token}` or header `x-prismbox-user-token:
      {token}`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"